+++ CHANGELOG.md
@@ -47,6 +47,9 @@ This release introduces `configVersion: 2` with clearer rules for the configurat
   * It is still suggested to set `autoUpdateConfig` to `true` so users don't miss out on future new features.
 * **Improvement.** Better guard protection when multiple recordings appear in burst (none will process).
 * **Improvement.** `--insert, --get-insert, --exec-insert, getInsert, execInsert` flags have been deprecated for quite some time, and now they've been cleaned up from the code.
+* **Improvement.** Use the `segments` key from the `meta.json` file as placeholder, and the speaker-separated transcript will be automatically formatted correctly.
+  * It is now also possible to use `meta.json` keys from the API, keys, or arrays. For example, `{{promptContext.systemContext.language}}` is possible.
+
 ---
 ## [v1.4.0](https://github.com/ognistik/macrowhisper/releases/tag/v1.4.0) - 2026/02/13
 ### Breaking